+++ lib/vcfb/acts_as_form_label_component.rb
@@ -9,6 +9,7 @@ module VCFB
 
     included do
       delegate :translation, :to_s, to: :label_builder
+      alias_method :default_text, :translation
     end
 
     private