+++ lib/vcfb/template.rb
@@ -1,4 +1,9 @@
 module VCFB
+  # Several FormBuilder methods are more complex and have changed a good bit
+  # between between Rails releases. Rather than trying to keep up with these
+  # changes, we put together an intermediate Template class (see
+  # VCFB::FormBuilder) to defer the compnetifying until _after_ it's gone
+  # through ActionView::Helpers::FormBuilder.
   class Template
     def initialize(template, form)
       @template = template
@@ -7,14 +12,11 @@ module VCFB
 
     # Custom Tag Helpers
 
-    def button_tag(value, options, &block)
+    def button_tag(value, options)
       return super unless @form.component_defined?(:button)
 
-      if block
-        value = @template.capture(value, &block)
-      end
-
-      @form.componentify(:button, value, options)
+      block = options.delete(:_block_for_component)
+      @form.componentify_with_slots(:button, value, options, &block)
     end
 
     def submit_tag(value, options)