+++ spec/internal/app/components/slot_form/collection_radio_buttons/label/component.rb
@@ -1,8 +1,14 @@
-module Form
+module SlotForm
   module CollectionRadioButtons
     module Label
       class Component < ViewComponent::Base
         include VCFB::ActsAsFormLabelComponent
+
+        if Gem::Version.new(ViewComponent::VERSION::STRING) >= Gem::Version.new("2.54.0")
+          renders_one :icon, ->(name) do
+            content_tag(:i, nil, class: "fa fa-#{name}")
+          end
+        end
 
         def initialize(form, method, text = nil, options = {})
           @form = form