+++ spec/internal/app/components/slot_form/collection_radio_buttons/radio_button/component.rb
@@ -1,0 +1,20 @@
+module SlotForm
+  module CollectionRadioButtons
+    module RadioButton
+      class Component < ViewComponent::Base
+        include VCFB::ActsAsFormComponent
+
+        def initialize(form, method, value, options = {})
+          @form = form
+          @method = method
+          @value = value
+          @options = options
+        end
+
+        def form_element(options = {})
+          radio_button(@form.object_name, @method, @value, options)
+        end
+      end
+    end
+  end
+end