+++ spec/lib/vcfb/form_builder/button_spec.rb
@@ -4,4 +4,14 @@ RSpec.describe "VCFB::FormBuilder#button", type: :helper do
     "with only options" => [{value: true}],
     "when passed block" => [->(*) { "<strong>Accept EULA</strong>".html_safe }]
   }
+
+  if Gem::Version.new(ViewComponent::VERSION::STRING) >= Gem::Version.new("2.54.0")
+    it "renders slots" do
+      form = nil
+      helper.form_with(model: Author.new, builder: SlotFormBuilder) { |f| form = f }
+      result = form.button { |component| component.with_icon name: :taco }
+      expect(normalize_output(result))
+        .to eq '<button name="button" type="submit" class="test"><i class="fa fa-taco"></i>Create Author</button>'
+    end
+  end
 end