+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Main
 
 - Support rendering of the base form tag via component.
+- Support rendering of slots for button, label, collection checkbox label, and
+  collection radio button label components.
 
 > **Note**: If you've previously the component generator, you'll need to re-run
 > it and merge the changes into your existing components.